% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_find_definition.R
\name{gh_find_func_definition}
\alias{gh_find_func_definition}
\title{Find Function Definitions on GitHub}
\usage{
gh_find_func_definition(repo_name, func_name)
}
\arguments{
\item{repo_name}{a character as the repository name that \code{func_name} is defiened. If you don't pass a full repository name, it guesses.}

\item{func_name}{a character as a function name.}
}
\description{
Find Function Definitions on GitHub
}
\examples{
\dontrun{
gh_find_func_definition("dplyr", "mutate")
}

}

