% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_get_package_info.R
\name{gh_get_package_info}
\alias{gh_get_package_info}
\title{Get Information of Packages on GitHub}
\usage{
gh_get_package_info(authors = NULL)
}
\arguments{
\item{authors}{a character vector as GitHub username. If you set \code{NULL} (default), it returns all packages information.}
}
\value{
a data.frame that has author, package name and title.
}
\description{
Get Information of Packages on GitHub
}
\examples{
\dontrun{
gh_get_package_info("hadley")
}

}

