% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_guess.R
\name{gh_guess}
\alias{gh_guess}
\title{Guess Github Repository Name from a Incomplete Name}
\usage{
gh_guess(repo_name, keep_title = FALSE)
}
\arguments{
\item{repo_name}{a character. A part of a repository name.}
}
\value{
candidates for the repository name.
}
\description{
Guess Github Repository Name from a Incomplete Name
}
\examples{
gh_guess("AnomalyDetection")
# [1] "twitter/AnomalyDetection"
gh_guess("BnomalyDetection")
# [1] "twitter/AnomalyDetection"
gh_guess("Uwitter/BnomalyDetection")
# [1] "twitter/AnomalyDetection"

}

