% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_list_packages.R
\name{gh_list_packages}
\alias{gh_list_packages}
\title{Get Information of Packages on GitHub}
\usage{
gh_list_packages(username = NULL)
}
\arguments{
\item{username}{a character vector as GitHub username. If you set \code{NULL} (default), it returns all packages information.}
}
\value{
a data.frame that has author, package name and title.
}
\description{
Get Information of Packages on GitHub
}
\examples{
\dontrun{
gh_list_packages("hadley")
}

}
