% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_search_packages.R
\name{gh_search_packages}
\alias{gh_search_packages}
\title{Search Packages from Titles by Regular Expressions}
\usage{
gh_search_packages(regex, ignore.case = TRUE)
}
\arguments{
\item{regex}{a character string containing a \link{regular expression} to be matched in the package titles.}

\item{ignore.case}{logical. If \code{FALSE}, the pattern matching is case sensitive and if \code{TRUE}, case is ignored during matching.}
}
\value{
a data.frame of package information.
}
\description{
Search Packages from Titles by Regular Expressions
}
\examples{
\dontrun{
gh_search_packages("lasso")
}

}
