% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_suggest_username.R
\name{gh_suggest_username}
\alias{gh_suggest_username}
\title{Suggest Github Username from a Faint Memory}
\usage{
gh_suggest_username(vague_name)
}
\arguments{
\item{vague_name}{a character. GitHub username that may not be exact.}
}
\value{
a character vector of the closest usernames to input.
}
\description{
Suggest Github Username from a Faint Memory
}
\details{
The trouble is that the usernames of GitHub are often hard to remember.
The function provides a way to obtain usernames from a faint memory.
}
\examples{
\dontrun{
gh_guess_username("yuhui")
# [1] "yihui"
}

}

