% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_update_package_list.R
\name{gh_update_package_list}
\alias{gh_update_package_list}
\title{Update the List of Packages on GitHub.}
\usage{
gh_update_package_list()
}
\description{
Update the List of Packages on GitHub.
}
