% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_install_packages.R, R/githubinstall.R
\name{gh_install_packages}
\alias{gh_install_packages}
\alias{githubinstall}
\title{Install Packages from GitHub}
\usage{
gh_install_packages(pkgs, build_args = NULL, build_vignettes = TRUE,
  uninstall = FALSE, verbose = TRUE, dependencies = c("Depends",
  "Imports", "Suggests"), ...)

githubinstall(pkgs, build_args = NULL, build_vignettes = TRUE,
  uninstall = FALSE, verbose = TRUE, dependencies = c("Depends",
  "Imports", "Suggests"), ...)
}
\arguments{
\item{pkgs}{character vector of the names of packages.}

\item{build_args}{character string used to control the package build, passed to \code{R CMD build}.}

\item{build_vignettes}{logical specifying whether to build package vignettes, passed to \code{R CMD build}. Can be slow. Default is \code{TRUE}.}

\item{uninstall}{logical}

\item{verbose}{logical specifying whether to print details of package building and installation.}

\item{dependencies}{character vector specifying which dependencies to install (of "Depends", "Imports", "Suggests", etc.).}

\item{...}{additional arguments to control installation of package, passed to \link{install.packages}.}
}
\value{
A named character vector of versions of R packages installed.
}
\description{
Install Packages from GitHub
}
\details{
\code{githubinstall()} is an alias of \code{gh_install_packages()}.
}
\examples{
\dontrun{
githubinstall("multidplyr")
gh_install_packages("multidplyr")
}

}

