% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_for_install.R
\name{separate_into_package_and_reference}
\alias{separate_into_package_and_reference}
\title{The "repo" argument allows to contain "ref" as 
"package_name@ref", "package_name#pull" or "package_name[branch]".
The function detects that "repo" contains "ref" and separates into pure repo and ref.
If "repo" contains "ref" and "ref" argument is specified, the values in "repo" take precedence.}
\usage{
separate_into_package_and_reference(packages, original_ref)
}
\arguments{
\item{packages}{"repo" argument.}

\item{original_ref}{"ref" argument.}
}
\description{
The "repo" argument allows to contain "ref" as 
"package_name@ref", "package_name#pull" or "package_name[branch]".
The function detects that "repo" contains "ref" and separates into pure repo and ref.
If "repo" contains "ref" and "ref" argument is specified, the values in "repo" take precedence.
}
