% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_guess.R
\name{gh_guess}
\alias{gh_guess}
\title{Guess Github Repository from Incomplete or Vague Names}
\usage{
gh_guess(vague_repo_name)
}
\arguments{
\item{repo_name}{a character.}
}
\value{
complete repository name.
}
\description{
Guess Github Repository from Incomplete or Vague Names
}
\examples{
\dontrun{
gh_guess("bigquery")
# [1] "bigrquery"
gh_guess_repository("bigquery", fullname = TRUE)
# [1] "hadley/bigrquery"    "rstats-db/bigrquery"
gh_guess_repository("hadley/bigquery")
# [1] "hadley/bigrquery"
}

}

