% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_guess_repository.R
\name{gh_guess_repository}
\alias{gh_guess_repository}
\title{Guess Github Repository Name from a Faint Memory}
\usage{
gh_guess_repository(vague_repo_name, fullname = FALSE)
}
\arguments{
\item{vague_repo_name}{a character as GitHub repository name that may not be exact.
You can pass full repository names e.g. "hadley/bigquery".}

\item{fullname}{logical. Indicates to return full repoitory names e.g. "hadley/bigrquery". 
If you pass full repository name to \code{vague_repo_name}, force \code{TRUE}.
Default is \code{FALSE}.}
}
\value{
a character vector of the closest repository names to input.
}
\description{
Guess Github Repository Name from a Faint Memory
}
\examples{
\dontrun{
gh_guess_repository("bigquery")
# [1] "bigrquery"
gh_guess_repository("bigquery", fullname = TRUE)
# [1] "hadley/bigrquery"    "rstats-db/bigrquery"
gh_guess_repository("hadley/bigquery")
# [1] "hadley/bigrquery"
}

}

