% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_show_source.R
\name{gh_show_source}
\alias{gh_show_source}
\title{Find source code for functions in packages on GitHub}
\usage{
gh_show_source(func, repo = NULL)
}
\arguments{
\item{func}{a function or a character. A function name.}

\item{repo}{a character. A GitHub repository name that must not be exactry.}
}
\description{
Find source code for functions in packages on GitHub
}
\examples{
\dontrun{
gh_show_source("mutate", "dplyr")

library(dplyr)
gh_show_source(mutate)
}

}

