% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_sugget.R
\name{gh_suggest}
\alias{gh_suggest}
\title{Suggest Github Repository from a Incomplete Name}
\usage{
gh_suggest(repo_name, keep_title = FALSE)
}
\arguments{
\item{repo_name}{a character. A part of a repository name.}

\item{keep_title}{logical. Indicates to keep the package titles as an attrbite. Default \code{FALSE}.}
}
\value{
candidates for the repository name.
}
\description{
Suggest Github Repository from a Incomplete Name
}
\examples{
gh_suggest("AnomalyDetection")
# [1] "twitter/AnomalyDetection"
gh_suggest("BnomalyDetection")
# [1] "twitter/AnomalyDetection"
gh_suggest("uwitter/BnomalyDetection")
# [1] "twitter/AnomalyDetection"

}

