% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_install_packages.R, R/githubinstall.R
\name{gh_install_packages}
\alias{gh_install_packages}
\alias{githubinstall}
\title{Install Packages from GitHub}
\usage{
gh_install_packages(packages, ask = TRUE, build_args = NULL,
  build_vignettes = FALSE, verbose = TRUE, dependencies = c("Depends",
  "Imports", "LinkingTo"), ...)

githubinstall(packages, ask = TRUE, build_args = NULL,
  build_vignettes = FALSE, verbose = TRUE, dependencies = c("Depends",
  "Imports", "LinkingTo"), ...)
}
\arguments{
\item{packages}{character vector of the names of packages.}

\item{ask}{logical. Indicates ask to confirm before install.}

\item{build_args}{character string used to control the package build, passed to \code{R CMD build}.}

\item{build_vignettes}{logical specifying whether to build package vignettes, passed to \code{R CMD build}. Can be slow. Default is \code{FALSE}.}

\item{verbose}{logical specifying whether to print details of package building and installation.}

\item{dependencies}{character vector specifying which dependencies to install (of "Depends", "Imports", "Suggests", etc.).}

\item{...}{additional arguments to control installation of package, passed to \link{install.packages}.}
}
\value{
A named character vector of versions of R packages installed.
}
\description{
Install Packages from GitHub
}
\details{
\code{githubinstall()} is an alias of \code{gh_install_packages()}.
}
\examples{
\dontrun{
githubinstall("AnomalyDetection")
gh_install_packages("AnomalyDetection")
}

}

